module rename to 'theme', delete 'TestThemeProvider.tsx'
