+++ apps/src/tests/issue-tests/index.ts
@@ -205,4 +205,3 @@ export { default as TestScreenStack } from './TestScreenStack';
 export { default as TestSplit } from './TestSplit';
 export { default as TestSafeAreaViewIOS } from './TestSafeAreaViewIOS';
 export { default as TestStackNesting } from './TestStackNesting';
-export { default as TestThemeProvider } from './TestThemeProvider';
